[CHORE] Use new reward interface for chest rewards
